refactor: split large files (server.go 2076→11 files, ScriptList.tsx 1047→4 files)
Split internal/agentmanager/server.go into domain-specific handler files
(task, interaction, agent sync, worktree, git, script, conflict resolution,
permissions) and ScriptList.tsx into utility/hook/component files.
No logic changes - pure file reorganization.
